requirements.txt

PyYAML==6.0.1
Requests==2.31.0

